Change visibility for struct fields to pub
In order to be able to use the API people need to be able to access the
data or construct input for the API requests. However, while enums being
public make every field visible this is not the case for structs. Until
convenience methods for construction or working with the data is
created, we've defaulted all struct fields to public so others can kick
the tires with the API even before we release.