Fix time bugs and serialization errors
We have two issues that arose while testing the API (not all of it is
tested yet). The first issue was time based. While most of the docs
say they will use ISO 8601 timestamps it doesn't guarantee a format.
While we thought it'd be a Date Time object with no offset (Zulu Time),
we found in some cases you might just get a date with no timestamp. To
account for this we created an enum to handle both cases. The second
error was that the API expected values to be undefined and not null if
being used as input. We fix this by skipping it's serialization if the
value is None.